Fix #47 - Removed `Evaluate script` button
Given how things work now, the evaluation on save button is rather
useless and it clutters the UI, hence the decision of removing it.

A worksheet is re-evaluated every time the user presses save (Re #3).
If the user wants to be in control, it can disable re-evaluation on save
(Re #46) and simply use the keyboard shortcut CMD/Ctrl+SHIFT+B.
Either way, if there is any error, the Problem view will report them.