Fix #57, Fix #58 - Stop evaluation on editor's  `dispose`
The current solution seem to be working, but it's not thread-safe (meaning
that it is still possible to get a NPE, just very unlikely). That said, the code
is robust enough and can recover from it, i.e., the worksheet will still be
functional even if a NPE is thrown.

I've added a big comment on the reason why the current solution is not
thread-safe, we'll have to see how to fix this.