Major clean-up in POM's properties
This clean-up was motivated by the impossibility of building the worksheet
against the freshly staged 2.1-M3 for Juno. In short, the wrong ``eclipse.repo`` and
``ajdt.repo`` where provided when building the worksheet for Juno.

With the intent of making the POM easier to understand, I've created a number of new
profiles that should simplify the task of building a release across different Scala
and Eclipse versions.

Last, I took the opportunity to also move to the latest released Tycho, i.e., 0.16.

I'll make sure to update the wiki to reflect the changes.
(cherry picked from commit c3349c19644bae3a86739bd6c11c1bbe4b80c190)