Allow multibyte characters in a worksheet
These changes attempt to fix an issue where a worksheet would be instrumented using the platform's default encoding rather than the project's. The fix collects its encoding by calling getDefaultCharset() on the eclipse project resource.

Be aware that the fix is ignorant to file-level encoding and/or project encoding changes; changing the worksheet's encoding will cause it to fail.

Fix #120