More robust WorksheetsManager: don't die on unhandled exceptions.
I noticed the worksheet manager actor died on some uncaught exceptions. This
commit makes it log and continue (since it's a singleton instance in a lazy val,
the only workaround is to restart eclipse). Also, noticed that closing a Scala
project causes `unit.scalaProject` to throw, so wrapped that into a `Try`.

I didn't generalize this treatment because in most other cases there is no
good way to recover from missing an open project.

Fixed #129