No need to watch actor's children as we can rely on supervision
Akka actors have supervision built-in, hence there is non need for
explicitly monitoring termination of children as we use to do when using
Scala actors.

Akka default supervision strategy will take care of reinitializing
`WorksheetRunner` actors if a failure occurs. The related `ActorRef` that
we stored in the `worksheets` Map will always have a pointer to the alive
actor (i.e., reinitialization of an actor upon failure is transparent wrt
its `ActorRef`).